refactor(checker): route jsdoc import constraint relation guard
