+++ crates/tsz-checker/src/jsdoc/mod.rs
@@ -36,6 +36,7 @@
 //! - New data structures → `types.rs`
 
 pub(crate) mod diagnostics;
+pub(crate) mod diagnostics_import_type_constraints;
 pub(crate) mod diagnostics_imports;
 pub(crate) mod diagnostics_templates;
 pub(crate) mod lookup;
